All Clear Sign Hitters powered by motor, wheels have mountain holding power

+++ FtcRobotController/src/main/java/com/qualcomm/ftcrobotcontroller/opmodes/OpMode4102.java
@@ -1,37 +1,4 @@
-/* Copyright (c) 2014 Qualcomm Technologies Inc
-
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without modification,
-are permitted (subject to the limitations in the disclaimer below) provided that
-the following conditions are met:
-
-Redistributions of source code must retain the above copyright notice, this list
-of conditions and the following disclaimer.
-
-Redistributions in binary form must reproduce the above copyright notice, this
-list of conditions and the following disclaimer in the documentation and/or
-other materials provided with the distribution.
-
-Neither the name of Qualcomm Technologies Inc nor the names of its contributors
-may be used to endorse or promote products derived from this software without
-specific prior written permission.
-
-NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS
-LICENSE. THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
-THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
-FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
-
 package com.qualcomm.ftcrobotcontroller.opmodes;
-
-
 
 import com.qualcomm.ftcrobotcontroller.systems.AllClearSignHitters;
 import com.qualcomm.ftcrobotcontroller.systems.Arm;
@@ -43,37 +10,34 @@ import com.qualcomm.ftcrobotcontroller.systems.Necessities;
 import com.qualcomm.ftcrobotcontroller.systems.Wheels;
 import com.qualcomm.ftcrobotcontroller.systems.ZipLineHitters;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.HardwareMap;
+
+/**
+ * Created by benorgera on 11/1/15.
+ */
 
 public class OpMode4102 extends LinearOpMode {
 
 
-    //all toggle button controls below
-    int toggleThreashold = 30;
+    //toggle button controls
 
-    int toggleCount = 0;
-    int lastSwing = 0;
-    int lastDrop = 0;
-    int lastZip = 0;
-    int lastZip2 = 0;
-    int lastLatch = 0;
-    int lastHammerR = 0;
-    int lastHammerL = 0;
+    boolean lastSwingState = false;
+    boolean lastDropState = false;
+    boolean lastZipState = false;
+    boolean lastZip2State = false;
+    boolean lastLatchState = false;
 
-    private Arm arm;
-    private ButtonPusher buttonPusher;
-    private ClimberDepositor climberDepositor;
-    private ZipLineHitters zipLineHitters;
-    private Intake intake;
-    private Wheels wheels;
-    private AllClearSignHitters allClearSignHitters;
 
-    private drivingMode mode = OpMode4102.drivingMode.NoAngle;
+    //robot systems
 
-    private final double maxDcMotorPower = 1.0;
-    private final double minDcMotorPower = 0.0;
+    Arm arm;
+    ButtonPusher buttonPusher;
+    ClimberDepositor climberDepositor;
+    ZipLineHitters zipLineHitters;
+    Intake intake;
+    Wheels wheels;
+    AllClearSignHitters allClearSignHitters;
 
-    private Necessities n;
+    Necessities n;
 
     public void initialize() {
 
@@ -87,7 +51,7 @@ public class OpMode4102 extends LinearOpMode {
 
         arm = new Arm(hardwareMap.servo.get("Latch"), hardwareMap.dcMotor.get("Arm"), n);
 
-        allClearSignHitters = new AllClearSignHitters(hardwareMap.servo.get("AllClearL"), hardwareMap.servo.get("AllClearR"), n);
+        allClearSignHitters = new AllClearSignHitters(hardwareMap.dcMotor.get("AllClear"), n);
 
         climberDepositor = new ClimberDepositor(hardwareMap.servo.get("PL Arm"), hardwareMap.servo.get("PL Drop"), n);
 
@@ -101,21 +65,26 @@ public class OpMode4102 extends LinearOpMode {
 
         while (opModeIsActive()) {
 
-            toggleCount++;
 
             wheels.processMainstream(gamepad1.left_stick_x, gamepad1.left_stick_y * -1);
 
-            if (gamepad1.a && (toggleCount - lastSwing >= toggleThreashold)) { //toggle
-                climberDepositor.swing();
-                lastSwing = toggleCount;
 
-            }
+            boolean temp = gamepad1.a; //use temp variable just in case button state changes mid-operation
 
-            if (gamepad1.b && (toggleCount - lastDrop >= toggleThreashold)) { //toggle
-                climberDepositor.drop();
-                lastDrop = toggleCount;
+            if (temp && !lastSwingState) climberDepositor.swing(); //toggle
 
-            }
+            lastSwingState = temp;
+
+
+            temp = gamepad1.b;
+
+            if (temp && !lastDropState) climberDepositor.drop(); //toggle
+
+            lastDropState = temp;
+
+            if (gamepad1.left_bumper) wheels.toggleHolding(false);
+
+            if (gamepad1.right_bumper) wheels.toggleHolding(true);
 
             if (gamepad1.dpad_down) wheels.toggleBackward(true);
 
@@ -130,43 +99,45 @@ public class OpMode4102 extends LinearOpMode {
             }
 
 
+            //all clear sign hitters
 
+            if (gamepad2.y) {
 
-            if (gamepad2.x && (toggleCount - lastLatch >= toggleThreashold)) { //toggle
-                arm.latch();
-                lastLatch = toggleCount;
+                allClearSignHitters.swing();
+
+            } else if (gamepad2.dpad_up) {
+
+                allClearSignHitters.raise();
+
+            } else if (gamepad2.dpad_down) {
+
+                allClearSignHitters.lower();
+
+            } else {
+
+                allClearSignHitters.neutralize();
+
             }
 
-            if (gamepad2.left_bumper && (toggleCount - lastZip >= toggleThreashold)) { //toggle
-                zipLineHitters.toggleLeft();
-                lastZip = toggleCount;
-            }
+            temp = gamepad2.x;
 
-            if (gamepad2.right_bumper && (toggleCount - lastZip2 >= toggleThreashold)) { //toggle
-                zipLineHitters.toggleRight();
-                lastZip2 = toggleCount;
-            }
+            if (temp && !lastLatchState) arm.latch(); //toggle
 
-            if (gamepad2.dpad_up) allClearSignHitters.raise();
-
-            if (gamepad2.dpad_down) allClearSignHitters.lower();
-
-            if (gamepad2.dpad_left && toggleCount - lastHammerL >= toggleThreashold) {
-                allClearSignHitters.toggleLeft();
-                lastHammerL = toggleCount;
-            }
-
-            if (allClearSignHitters.isHammeringLeft()) allClearSignHitters.hammerLeft(toggleCount);
-
-            if (gamepad2.dpad_right && toggleCount - lastHammerR >= toggleThreashold) {
-                allClearSignHitters.toggleRight();
-                lastHammerR = toggleCount;
-            }
-
-            if (allClearSignHitters.isHammeringRight()) allClearSignHitters.hammerRight(toggleCount);
+            lastLatchState = temp;
 
 
-            waitOneFullHardwareCycle();
+            temp = gamepad2.left_bumper;
+
+            if (temp && !lastZipState) zipLineHitters.toggleLeft(); //toggle
+
+            lastZipState = temp;
+
+
+            temp = gamepad2.right_bumper;
+
+            if (temp && !lastZip2State) zipLineHitters.toggleRight(); //toggle
+
+            lastZip2State = temp;
         }
 
     }
@@ -181,11 +152,6 @@ public class OpMode4102 extends LinearOpMode {
 
         run();
 
-    }
-
-
-    public enum drivingMode {
-        NoAngle, Angle
     }
 
 }
